Move storage mock to a separate file
