+++ services/blockstorage/sync/harness.go
@@ -17,44 +17,6 @@ import (
 	"time"
 )
 
-// the storage mock should be moved to its own file, but i have a weird goland bug where it will not identify it and its driving me mad, putting this here for now
-type blockSyncStorageMock struct {
-	mock.Mock
-}
-
-func (s *blockSyncStorageMock) GetLastCommittedBlockHeight(ctx context.Context, input *services.GetLastCommittedBlockHeightInput) (*services.GetLastCommittedBlockHeightOutput, error) {
-	ret := s.Called(ctx, input)
-	if out := ret.Get(0); out != nil {
-		return out.(*services.GetLastCommittedBlockHeightOutput), ret.Error(1)
-	} else {
-		return nil, ret.Error(1)
-	}
-}
-
-func (s *blockSyncStorageMock) CommitBlock(ctx context.Context, input *services.CommitBlockInput) (*services.CommitBlockOutput, error) {
-	ret := s.Called(ctx, input)
-	if out := ret.Get(0); out != nil {
-		return out.(*services.CommitBlockOutput), ret.Error(1)
-	} else {
-		return nil, ret.Error(1)
-	}
-}
-
-func (s *blockSyncStorageMock) ValidateBlockForCommit(ctx context.Context, input *services.ValidateBlockForCommitInput) (*services.ValidateBlockForCommitOutput, error) {
-	ret := s.Called(ctx, input)
-	if out := ret.Get(0); out != nil {
-		return out.(*services.ValidateBlockForCommitOutput), ret.Error(1)
-	} else {
-		return nil, ret.Error(1)
-	}
-}
-
-func (s *blockSyncStorageMock) UpdateConsensusAlgosAboutLatestCommittedBlock(ctx context.Context) {
-	s.Called(ctx)
-}
-
-// end of storage mock
-
 type blockSyncConfigForTests struct {
 	pk               primitives.Ed25519PublicKey
 	batchSize        uint32
